requirements.txt
# requirements.txt
streamlit
pillow
pandas
matplotlib
python-dotenv
requests
google-genai==1.7.0
fuzzywuzzy 
python-Levenshtein